+++ changelog.md
@@ -1,9 +1,17 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 ## Update
+## [0.1.5] - 2020-04-20
+* Get and show posts by category.
+* Readme updated.
+## Update
+## [0.1.4] - 2020-04-20
+* View for COVID-19 statistics with charts.
+## Update
 ## [0.1.3] - 2020-04-19
 * News blog, add new post and user profile read only component added.
 * Service to retrieve all posts and create new post.
+* Logout function.
 ## Update
 ## [0.1.2] - 2020-04-18
 * Login, register and profile views and components.